Move around machine architecture testers
