--- subprojects/runtime-native/src/test/groovy/dev/nokee/runtime/nativebase/CommonMachineArchitectureTester.java
@@ -1,9 +1,0 @@
-package dev.nokee.runtime.nativebase;
-
-public interface CommonMachineArchitectureTester extends MachineArchitectureItaniumTester,
-	MachineArchitecturePowerPC32BitTester, MachineArchitecturePowerPC64BitTester,
-	MachineArchitectureIntel32BitTester, MachineArchitectureIntel64BitTester,
-	MachineArchitectureSparc32BitTester, MachineArchitectureSparc64BitTester,
-	MachineArchitectureHPPARISCTester,
-	MachineArchitectureUnknownTester
-{}
