+++ subprojects/runtime-native/src/test/groovy/dev/nokee/runtime/nativebase/KnownMachineArchitecturePaRiscTester.java
@@ -1,24 +1,20 @@
 package dev.nokee.runtime.nativebase;
 
-import dev.nokee.runtime.nativebase.MachineArchitecture;
-import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import static dev.nokee.internal.testing.GradleNamedMatchers.named;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-/** @see dev.nokee.runtime.nativebase.CommonMachineArchitectureTester */
-interface MachineArchitectureItaniumTester {
+/** @see KnownMachineArchitectureTester */
+interface KnownMachineArchitecturePaRiscTester {
 	MachineArchitecture createSubject(String name);
 
 	@ParameterizedTest(name = "has 64-bit pointer size [{arguments}]")
-	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonItaniumNames")
-	default void itaniumArchitectureHas64BitPointerSize(String name) {
+	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonHPPARISCNames")
+	default void hpPARISCArchitectureHas64BitPointerSize(String name) {
 		assertAll(
 			() -> assertThat(createSubject(name).is32Bit(), is(false)),
 			() -> assertThat(createSubject(name).is64Bit(), is(true))
@@ -26,8 +22,8 @@ interface MachineArchitectureItaniumTester {
 	}
 
 	@ParameterizedTest(name = "has canonical name [{arguments}]")
-	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonItaniumNames")
-	default void itaniumArchitectureHasCanonicalName(String name) {
-		assertThat(createSubject(name), named("Itanium"));
+	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonHPPARISCNames")
+	default void hpPARISCArchitectureHasCanonicalName(String name) {
+		assertThat(createSubject(name), named("PA-RISC"));
 	}
 }
