+++ subprojects/runtime-native/src/test/groovy/dev/nokee/runtime/nativebase/KnownMachineArchitectureSparc32BitTester.java
@@ -1,6 +1,5 @@
 package dev.nokee.runtime.nativebase;
 
-import dev.nokee.runtime.nativebase.MachineArchitecture;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
@@ -9,13 +8,13 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-/** @see dev.nokee.runtime.nativebase.CommonMachineArchitectureTester */
-interface MachineArchitectureIntel32BitTester {
+/** @see KnownMachineArchitectureTester */
+interface KnownMachineArchitectureSparc32BitTester {
 	MachineArchitecture createSubject(String name);
 
 	@ParameterizedTest(name = "has 32-bit pointer size [{arguments}]")
-	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonIntel32BitNames")
-	default void intel32BitArchitectureHas32BitPointerSize(String name) {
+	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonSparc32BitNames")
+	default void sparc32BitArchitectureHas32BitPointerSize(String name) {
 		assertAll(
 			() -> assertThat(createSubject(name).is32Bit(), is(true)),
 			() -> assertThat(createSubject(name).is64Bit(), is(false))
@@ -23,8 +22,8 @@ interface MachineArchitectureIntel32BitTester {
 	}
 
 	@ParameterizedTest(name = "has canonical name [{arguments}]")
-	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonIntel32BitNames")
-	default void intel32BitArchitectureHasCanonicalName(String name) {
-		assertThat(createSubject(name), named("x86"));
+	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonSparc32BitNames")
+	default void sparc32BitArchitectureHasCanonicalName(String name) {
+		assertThat(createSubject(name), named("Sparc"));
 	}
 }
