+++ subprojects/runtime-native/src/test/groovy/dev/nokee/runtime/nativebase/KnownMachineArchitectureSparc64BitTester.java
@@ -8,13 +8,13 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-/** @see dev.nokee.runtime.nativebase.CommonMachineArchitectureTester */
-interface MachineArchitecturePowerPC64BitTester {
+/** @see KnownMachineArchitectureTester */
+interface KnownMachineArchitectureSparc64BitTester {
 	MachineArchitecture createSubject(String name);
 
 	@ParameterizedTest(name = "has 64-bit pointer size [{arguments}]")
-	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonPowerPC64BitNames")
-	default void powerPC64BitArchitectureHas64BitPointerSize(String name) {
+	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonSparc64BitNames")
+	default void sparc64BitArchitectureHas64BitPointerSize(String name) {
 		assertAll(
 			() -> assertThat(createSubject(name).is32Bit(), is(false)),
 			() -> assertThat(createSubject(name).is64Bit(), is(true))
@@ -22,8 +22,8 @@ interface MachineArchitecturePowerPC64BitTester {
 	}
 
 	@ParameterizedTest(name = "has canonical name [{arguments}]")
-	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonPowerPC64BitNames")
-	default void powerPC64BitArchitectureHasCanonicalName(String name) {
-		assertThat(createSubject(name), named("PowerPC64"));
+	@MethodSource("dev.nokee.runtime.nativebase.MachineArchitectureTestUtils#commonSparc64BitNames")
+	default void sparc64BitArchitectureHasCanonicalName(String name) {
+		assertThat(createSubject(name), named("Sparc64"));
 	}
 }
