+++ subprojects/runtime-native/src/test/groovy/dev/nokee/runtime/nativebase/KnownMachineArchitectureTester.java
@@ -1,0 +1,8 @@
+package dev.nokee.runtime.nativebase;
+
+public interface KnownMachineArchitectureTester extends KnownMachineArchitectureItaniumTester,
+	KnownMachineArchitecturePowerPc32BitTester, KnownMachineArchitecturePowerPc64BitTester,
+	KnownMachineArchitectureIntel32BitTester, KnownMachineArchitectureIntel64BitTester,
+	KnownMachineArchitectureSparc32BitTester, KnownMachineArchitectureSparc64BitTester,
+	KnownMachineArchitecturePaRiscTester
+{}
